[quant][refactor tests] Move test_quantized_module into test_quantize_eager_ptq
Summary:

Test Plan:
python test/test_quantization.py TestStaticQuantizedModule
python test/test_quantization.py TestDynamicQuantizedModule
Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
